Adicionei o metodo showIdGato que serve para mostrar as informações do gato cadastrado, tornei o programa mais humano no sentido de ele perguntar os dados do gato falando pelo seu nome.

+++ Animais/src/animal/animal.java
@@ -2,7 +2,7 @@ package animal;
 
 public abstract class animal {
 	public String nome;
-	public Float peso;
+	public double peso;
 	public int idade;
 	public String corPelo;
